requirements.txt
anyio==3.5.0
asgiref==3.5.1
astroid==2.9.3
asttokens==2.0.5
attrs==21.4.0
audioread==3.0.1
blender==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.1.3
clip==0.2.0
colorama==0.4.4
Cython==3.0.5
decorator==4.4.2
emoji==1.7.0
eth-hash==0.3.2
eth-typing==2.2.2
executing==0.8.3
fastapi==0.77.1
ffmpeg==1.4
filelock==3.8.0
Flask==2.1.2
gptj==3.0.9
h11==0.13.0
huggingface-hub==0.10.1
idna==3.3
imageio==2.14.0
imageio-ffmpeg==0.4.5
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jupyter-core==4.9.1
lazy-object-proxy==1.7.1
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.38.0
MarkupSafe==2.1.1
matchering==2.0.3
mccabe==0.6.1
mido==1.3.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numba==0.55.0
numpy==1.21.5
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==23.2
pandas==1.3.5
parso==0.8.3
pathspec==0.9.0
patsy==0.5.2
pedalboard==0.8.5
Pillow==9.0.0
platformdirs==3.11.0
pooch==1.8.0
proglog==0.1.9
psutil==5.9.0
pure-eval==0.2.1
pycparser==2.21
pydantic==1.9.0
pydub==0.25.1
pylint==2.12.2
pyparsing==3.0.6
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2021.3
pywin32==303
PyYAML==6.0
pyzmq==23.2.0
redshift==0.0.1
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.2.2
scikit-learn==1.3.2
scipy==1.7.3
simpleaudio==1.0.4
six==1.12.0
sniffio==1.2.0
soundfile==0.12.1
soxr==0.3.7
stack-data==0.3.0
starlette==0.19.1
statsmodels==0.13.1
sympy==1.12
threadpoolctl==3.2.0
toml==0.10.2
tomli==1.2.3
toolz==0.11.1
torch==1.12.0
tornado==6.1
touch==2020.12.3
tqdm==4.62.3
traitlets==5.1.1
twitter-application-only-auth==0.3.3
typing_extensions==4.8.0
urllib3==1.26.15
uvicorn==0.17.6
values==2020.12.3
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.13.3
zstandard==0.22.0
